Refactor: grouping entities by responsibilities, add some comments.
