+++ rework_experiments/lib/navigation/lib/navigation/external/README.md
@@ -1,0 +1,1 @@
+This folder contains entities responsible for external navigation, such as showing modal sheet.
