+++ rework_experiments/lib/navigation/lib/navigation/internal/wolt_modal_sheet_path.dart
@@ -1,4 +1,4 @@
-/// Class describing the path to navigate.
+/// Class describing the path to navigate inside modal sheet.
 /// An every instance of this class means a concrete intent to show a
 /// corresponding page for the [path] with passed [arguments] to it.
 class WoltModalSheetPath {
