+++ rework_experiments/lib/navigation/lib/theme/theme_resolver.dart
@@ -1,6 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:rework_experiments/navigation/lib/theme/wolt_modal_sheet_theme_data.dart';
 
+/// A widget that resolves the theme for the [WoltModalSheet] widget.
+/// Priority of resolving:
+/// Theme of the concrete modal sheet, see [WoltModalSheet.style] ->
+/// Global theme of modal sheet, see [WoltModalSheetTheme] ->
+/// fallback default scheme.
 class ThemeResolver extends StatefulWidget {
   final Widget child;
   final WoltModalSheetStyle? styleOverride;
